+++ app/src/app/admin/production/production-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ProductionStatusesComponent } from './production-statuses/production-statuses.component';
 import { MachinesComponent } from './machines/machines.component';
-import { ProductionComponent } from './production.component';
+import { WorkorderComponent } from './workorder.component';
 import { WorkorderboardComponent } from './workorderboard/workorderboard.component';
 import { BomComponent } from './bom/bom.component';
 
@@ -10,7 +10,7 @@ import { BomComponent } from './bom/bom.component';
 const routes: Routes = [
   {
     path :'',
-    component: ProductionComponent
+    component: WorkorderComponent
   },
   {
     path :'productionstatuses',
